client: (352ebe4) feat(refuel): modified refuel system to make it callable by all entity types, fixed carrying capacit
Modified refuel system to make it callable by all entity types as long as they are either owned by
you or is an Asteroid. This will allow refueling all ships by all other ships, allow transport of
harvested fuel by Harvesters to fuel carriers and allow Harvesting of fuel from Asteroids. We also
check if the source is Asteroid, then the destination has to be a Harvester, this will allow the
refuel system to be used as Harvest Fuel from Asteroid.