client: (772d89c) feat(attack): attack based on range and heading
We now only permit attack based on the range and the heading of the ship. You can only fire in an
arc in front of your nose. We use the previous position to determine the direction the ship is
facing and then determine how far the attack position is from this line of attack. Higher the level,
the greater your range and greater the shooting arc.